Implemented unwarp code
I did not write a filter itself (it could be reasonable, but not necessary)
waiting for the need of it.

Close #17

+++ Code/Examples/Model1Registration2Contour.cxx
@@ -85,7 +85,7 @@ int main(int argc, char *argv[]) {
 
 	typedef itk::VectorImageToImageAdaptor<double,3u>            VectorToImage;
 	typedef itk::VectorResampleImageFilter
-			<DeformationFieldType,DeformationFieldType,double>   DisplacementResamplerType;
+			<DeformationFieldType,DeformationFieldType,float>   DisplacementResamplerType;
 	typedef itk::BSplineInterpolateImageFunction
 			                <DeformationFieldType>               InterpolatorFunction;
 	typedef itk::DisplacementFieldTransform<float, 3u>           TransformType;
